# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A Twitter-like social media API built with Go",
        "title": "Forgotten API",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/clubs/{id}/reading": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "List all book assignments for a club.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Reading"
                ],
                "summary": "List Club Book Assignments",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Club ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.ClubAssignmentResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/clubs/{id}/reading/assign": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Assign a book to a club for reading.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Reading"
                ],
                "summary": "Assign Book to Club",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Club ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Assign Book Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.AssignBookRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.ClubAssignmentResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/clubs/{id}/reading/checkpoint": {
            "patch": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Update the reading checkpoint for a club's current book assignment.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Reading"
                ],
                "summary": "Update Club Reading Checkpoint",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Club ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update Club Checkpoint Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdateClubCheckpointRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ClubAssignmentResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/clubs/{id}/reading/complete": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Mark the current book assignment for a club as completed.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Reading"
                ],
                "summary": "Complete Club Book Assignment",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Club ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ClubAssignmentResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/users/{id}/reading": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "List all reading progress entries for a user.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Reading"
                ],
                "summary": "List User Reading Progress",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.UserBookProgressResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/users/{id}/reading/history": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Retrieve the reading history of a user.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Reading"
                ],
                "summary": "Get User Reading History",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.UserReadingHistoryItem"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/users/{id}/reading/sync": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Synchronize the reading statistics for a user.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Reading"
                ],
                "summary": "Sync User Reading Stats",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/users/{id}/readings": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Start reading a book by providing the book ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Reading"
                ],
                "summary": "Start Reading a Book",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Start Reading Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.StartReadingRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.UserBookProgressResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/users/{id}/readings/{bookID}": {
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Update the reading progress of a book.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Reading"
                ],
                "summary": "Update Reading Progress",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Book ID",
                        "name": "bookID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update Reading Progress Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdateReadingProgressRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.UserBookProgressResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/users/{id}/readings/{bookID}/complete": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Mark a book as completed and optionally add a note.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Reading"
                ],
                "summary": "Complete Reading a Book",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Book ID",
                        "name": "bookID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Complete Reading Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CompleteReadingRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.UserBookProgressResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.AssignBookRequest": {
            "type": "object",
            "required": [
                "book_id"
            ],
            "properties": {
                "book_id": {
                    "type": "integer"
                },
                "checkpoint": {
                    "type": "string"
                },
                "due_date": {
                    "type": "string"
                },
                "start_date": {
                    "type": "string"
                },
                "target_page": {
                    "type": "integer"
                }
            }
        },
        "models.Book": {
            "type": "object",
            "properties": {
                "author": {
                    "type": "string"
                },
                "cover_url": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "genre": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "isbn": {
                    "type": "string"
                },
                "pages": {
                    "type": "integer"
                },
                "published_year": {
                    "type": "integer"
                },
                "rating": {
                    "type": "number",
                    "maximum": 5,
                    "minimum": 0
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.ClubAssignmentResponse": {
            "type": "object",
            "properties": {
                "book": {
                    "$ref": "#/definitions/models.Book"
                },
                "checkpoint": {
                    "type": "string"
                },
                "club_id": {
                    "type": "integer"
                },
                "due_date": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "start_date": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "target_page": {
                    "type": "integer"
                }
            }
        },
        "models.CompleteReadingRequest": {
            "type": "object",
            "properties": {
                "note": {
                    "type": "string",
                    "maxLength": 500
                }
            }
        },
...
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "models.ReadingLog": {
            "type": "object",
            "properties": {
                "assignment_id": {
                    "type": "integer"
                },
                "book_id": {
                    "type": "integer"
                },
                "club_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "from_page": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "minutes": {
                    "type": "integer"
                },
                "note": {
                    "type": "string"
                },
                "pages_delta": {
                    "type": "integer"
                },
                "to_page": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
        "models.StartReadingRequest": {
            "type": "object",
            "required": [
                "book_id"
            ],
            "properties": {
                "book_id": {
                    "type": "integer"
                }
            }
        },
        "models.SuccessResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "models.UpdateClubCheckpointRequest": {
            "type": "object",
            "properties": {
                "checkpoint": {
                    "type": "string"
                },
                "target_page": {
                    "type": "integer"
                }
            }
        },
...
        "models.UpdateReadingProgressRequest": {
            "type": "object",
            "properties": {
                "current_page": {
                    "type": "integer",
                    "minimum": 0
                },
                "minutes": {
                    "type": "integer",
                    "minimum": 0
                },
                "note": {
                    "type": "string",
                    "maxLength": 500
                },
                "pages_delta": {
                    "type": "integer"
                },
                "percent": {
                    "type": "number",
                    "maximum": 100,
                    "minimum": 0
                }
            }
        },
...
        "models.UserBookProgressResponse": {
            "type": "object",
            "properties": {
                "book_id": {
                    "type": "integer"
                },
                "current_page": {
                    "type": "integer"
                },
                "finished_at": {
                    "type": "string"
                },
                "percent": {
                    "type": "number"
                },
                "started_at": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "models.UserReadingHistoryItem": {
            "type": "object",
            "properties": {
                "book": {
                    "$ref": "#/definitions/models.Book"
                },
                "finished_at": {
                    "type": "string"
                },
                "logs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ReadingLog"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}